Add an interactive console executable
We are stubbing API requests for our test suite, but sometime we
need to interact with the actual API request and it's responses.
This commit adds an easier way to play with the actual API in
the console, lets start by running `bin/console`.

To make it more easier, it also provides us a `.sample.pryrc`
and all we need do is copy it as `.pryrc` and set your keys.