Extract class method creation behavior
Currently, we were creating class method for each of the action
of the resources. This limits us to use some of other facility
we could use, so this commit extract those behavior to a `Base`
class which will do the job to instantiating those object for
the class methods using some meta programming and we could use
the advantage of this extraction.