Remove resource related duplications
Currently, we were using the API communication related methods
in each of the model, but now we have a `Base` model to inherit
with. So let's remove resource request related duplication and
this will also make it easier in long run as we are extracting
the communication behavior in one place.