Extract `list` action to a module
Most of the case we are repeating the `list` action, it would be
ideal to put it on the `Base` but the problem is, all of the sub
class from base might not need to have `list`. So, based on the
scenario it's better to extract it to a module.