Fix some broken example in the readme
There are couple of example in the `readme` that are not correctly
formatted or using some hard coded value, which might create some
confusion.

This commit fixes those example and it also replace those hard
coded value with a variable name, so they developer will get the
idea on what they should provide there.