Remove `list` related spec duplications
The `.list` API supports one additional parameter, which can be
used to filter the listing, but we are repeating that in every
resources's `list` spec, which is kind of a duplication.

Now that we have moved `list` action to a module, so it makes
more sense to move those `list` specific edge cases a spec that
will test that module.