Fix `#initialize_http_header` related issue
In the `Veeqo::Request` we're using `#initialize_http_header` to
set API key and `Content-Type` header separately. But the way
`#initialize_http_header` work, every call to this method resets
the header and reinitialize with provided hash's key value pair

This was causing not to set the headers properly, last call to
the `#initialize_http_header` was overriding any other headers.

This commit fixes this, Now all headers should have been setted
up properly and work like a charm :)

Ref: https://github.com/ruby/ruby/blob/trunk/lib/net/http/header.rb#L14